+++ src/email/message/command/edit.rs
@@ -10,40 +10,48 @@ use pimalaya_tui::{
 use tracing::info;
 
 use crate::{
-    account::arg::name::AccountNameFlag,
-    config::TomlConfig,
-    envelope::arg::ids::EnvelopeIdArg,
+    account::arg::name::AccountNameFlag, config::TomlConfig, envelope::arg::ids::EnvelopeIdArg,
     folder::arg::name::FolderNameOptionalFlag,
-    message::arg::{body::MessageRawBodyArg, header::HeaderRawArgs},
 };
 
-/// Forward a message.
+/// Edit a message.
 ///
-/// This command allows you to forward the given message using the
+/// This command allows you to edit the given message using the
 /// editor defined in your environment variable $EDITOR. When the
 /// edition process finishes, you can choose between saving or sending
 /// the final message.
 #[derive(Debug, Parser)]
-pub struct MessageForwardCommand {
+pub struct MessageEditCommand {
     #[command(flatten)]
     pub folder: FolderNameOptionalFlag,
 
     #[command(flatten)]
     pub envelope: EnvelopeIdArg,
 
-    #[command(flatten)]
-    pub headers: HeaderRawArgs,
+    /// List of headers that should be visible at the top of the
+    /// message.
+    ///
+    /// If a given header is not found in the message, it will not be
+    /// visible. If no header is given, defaults to the one set up in
+    /// your TOML configuration file.
+    #[arg(long = "header", short = 'H', value_name = "NAME")]
+    pub headers: Vec<String>,
 
-    #[command(flatten)]
-    pub body: MessageRawBodyArg,
+    /// Edit the message on place.
+    ///
+    /// If set, the original message being edited will be removed at
+    /// the end of the command. Useful when you need, for example, to
+    /// edit a draft, send it then remove it from the Drafts folder.
+    #[arg(long, short = 'p')]
+    pub on_place: bool,
 
     #[command(flatten)]
     pub account: AccountNameFlag,
 }
 
-impl MessageForwardCommand {
+impl MessageEditCommand {
     pub async fn execute(self, printer: &mut impl Printer, config: &TomlConfig) -> Result<()> {
-        info!("executing forward message command");
+        info!("executing edit message command");
 
         let folder = &self.folder.name;
 
@@ -63,6 +71,7 @@ impl MessageForwardCommand {
                     .without_features()
                     .with_add_message(BackendFeatureSource::Context)
                     .with_send_message(BackendFeatureSource::Context)
+                    .with_delete_messages(BackendFeatureSource::Context)
             },
         )
         .build()
@@ -74,11 +83,21 @@ impl MessageForwardCommand {
             .await?
             .first()
             .ok_or(eyre!("cannot find message"))?
-            .to_forward_tpl_builder(account_config.clone())
-            .with_headers(self.headers.raw)
-            .with_body(self.body.raw())
-            .build()
+            .to_read_tpl(&account_config, |mut tpl| {
+                if !self.headers.is_empty() {
+                    tpl = tpl.with_show_only_headers(&self.headers);
+                }
+
+                tpl
+            })
             .await?;
-        editor::edit_tpl_with_editor(account_config, printer, &backend, tpl).await
+
+        editor::edit_tpl_with_editor(account_config, printer, &backend, tpl).await?;
+
+        if self.on_place {
+            backend.delete_messages(folder, &[id]).await?;
+        }
+
+        Ok(())
     }
 }